refactor: PostgreSQL重複エラー判定を共通化
PostgreSQL unique_violation のコードと判定処理を web 共通 util に集約した。

budget 作成エラー処理とテストは共通 util を直接参照するようにした。
