+++ apps/web/src/features/budgets/createCategoryBudget/categoryBudgetCreateError.ts
@@ -1,5 +1,4 @@
-// PostgreSQL SQLSTATE 23505 means unique_violation.
-export const POSTGRES_UNIQUE_VIOLATION_CODE = "23505"
+import { isPostgresUniqueViolationError } from "../../../utils/postgresError"
 
 const DUPLICATE_CATEGORY_BUDGET_MESSAGE =
   "A category budget for this category and month already exists."
@@ -11,12 +10,4 @@ export function toCategoryBudgetCreateErrorMessage(error: unknown): string {
   }
 
   return GENERIC_CREATE_CATEGORY_BUDGET_MESSAGE
-}
-
-function isPostgresUniqueViolationError(error: unknown): boolean {
-  if (!error || typeof error !== "object" || !("code" in error)) {
-    return false
-  }
-
-  return error.code === POSTGRES_UNIQUE_VIOLATION_CODE
 }
