+++ apps/web/src/features/budgets/createMonthlyBudget/CreateMonthlyBudgetModal/CreateMonthlyBudgetModal.test.tsx
@@ -4,8 +4,8 @@ import { afterEach, describe, expect, test, vi } from "vite-plus/test"
 
 import { server } from "../../../../test/msw/server"
 import { act, fireEvent, render, screen, waitFor, within } from "../../../../test/test-utils"
+import { POSTGRES_UNIQUE_VIOLATION_CODE } from "../../../../utils/postgresError"
 import { fillCreateMonthlyBudgetForm } from "../../test/utils/budgetCreationForm"
-import { POSTGRES_UNIQUE_VIOLATION_CODE } from "../monthlyBudgetCreateError"
 import * as stories from "./CreateMonthlyBudgetModal.stories"
 
 const { Default } = composeStories(stories)
