+++ apps/web/src/features/budgets/createMonthlyBudget/monthlyBudgetCreateError.ts
@@ -1,5 +1,4 @@
-// PostgreSQL SQLSTATE 23505 means unique_violation.
-export const POSTGRES_UNIQUE_VIOLATION_CODE = "23505"
+import { isPostgresUniqueViolationError } from "../../../utils/postgresError"
 
 const DUPLICATE_MONTHLY_BUDGET_MESSAGE = "A monthly budget for this month already exists."
 const GENERIC_CREATE_MONTHLY_BUDGET_MESSAGE = "Failed to create monthly budget."
@@ -10,12 +9,4 @@ export function toMonthlyBudgetCreateErrorMessage(error: unknown): string {
   }
 
   return GENERIC_CREATE_MONTHLY_BUDGET_MESSAGE
-}
-
-function isPostgresUniqueViolationError(error: unknown): boolean {
-  if (!error || typeof error !== "object" || !("code" in error)) {
-    return false
-  }
-
-  return error.code === POSTGRES_UNIQUE_VIOLATION_CODE
 }
